fix new lints
staticcheck has a new lint, QF1008, to remove superfluous references to
embedded fields. In this case, changing e.g. `pr.ObjectMeta.Annotations`
to `pr.Annotations` feels weird because it does not match the YAML
structure, so this commit rewrites this in a way that is still clear
while also making the linter happy.